chore: reorganize code structure and cleanup files
